Revert unnecessary changes in tests

+++ tests/epyccel/modules/arrays.py
@@ -561,18 +561,18 @@ def array_int32_in_bool_out_2d_F_complex_3d_expr(x : 'int32[:,:](order=F)', y : 
 
 @stack_array('a')
 def array_float_1d_sum_stack_array():
-    from numpy import zeros, float64
+    from numpy import zeros
     a = zeros(10)
-    s = float64(0.)
+    s = 0.
     for i in range(10):
         s += a[i]
     return s
 
 @stack_array('a')
 def array_float_1d_div_stack_array():
-    from numpy import ones, float64
+    from numpy import ones
     a = ones(10)
-    s = float64(0.)
+    s = 0.
     for i in range(10):
         s += 1.0 / a[i]
     return s
@@ -580,10 +580,10 @@ def array_float_1d_div_stack_array():
 @stack_array('a')
 @stack_array('b')
 def multiple_stack_array_1():
-    from numpy import ones, array, float64
+    from numpy import ones, array
     a = ones(5)
     b = array([1, 3, 5, 7, 9])
-    s = float64(0.0)
+    s = 0.0
     for i in range(5):
         s += a[i] / b[i]
     return s
@@ -591,11 +591,11 @@ def multiple_stack_array_1():
 @stack_array('a')
 @stack_array('b', 'c')
 def multiple_stack_array_2():
-    from numpy import ones, array, float64
+    from numpy import ones, array
     a = ones(5)
     b = array([2, 4, 6, 8, 10])
     c = array([1, 3, 5, 7, 9])
-    s = float64(0.0)
+    s = 0.0
     for i in range(5):
         s = s + b[i] - a[i] / c[i]
     return s
@@ -606,9 +606,9 @@ def multiple_stack_array_2():
 
 @stack_array('a')
 def array_float_2d_sum_stack_array():
-    from numpy import zeros, float64
+    from numpy import zeros
     a = zeros((10, 10))
-    s = float64(0.)
+    s = 0.
     for i in range(10):
         for j in range(10):
             s += a[i][j]
@@ -616,9 +616,9 @@ def array_float_2d_sum_stack_array():
 
 @stack_array('a')
 def array_float_2d_div_stack_array():
-    from numpy import full, float64
+    from numpy import full
     a = full((10, 10), 2)
-    s = float64(1.)
+    s = 1.
     for i in range(10):
         for j in range(10):
             s /= a[i][j]
@@ -627,10 +627,10 @@ def array_float_2d_div_stack_array():
 @stack_array('a')
 @stack_array('b')
 def multiple_2d_stack_array_1():
-    from numpy import ones, array, float64
+    from numpy import ones, array
     a = ones((2, 5))
     b = array([[1, 3, 5, 7, 9], [11, 13, 17, 19, 23]])
-    s = float64(0.0)
+    s = 0.0
     j = 0
     for i in range(2):
         for j in range(5):
@@ -640,11 +640,11 @@ def multiple_2d_stack_array_1():
 @stack_array('a')
 @stack_array('b', 'c')
 def multiple_2d_stack_array_2():
-    from numpy import ones, array, float64
+    from numpy import ones, array
     a = ones(5)
     b = array([[2, 4, 6, 8, 10], [1, 3, 5, 7, 9]])
     c = array([[1, 3, 5, 7, 9], [2, 4, 6, 8, 10]])
-    s = float64(0.0)
+    s = 0.0
     for i in range(2):
         for j in range(5):
             s = s + b[i][j] - a[j] / c[i][j]
@@ -1721,7 +1721,7 @@ def arr_arange_7(arr : 'int[:,:]'):
 def iterate_slice(i : int):
     import numpy as np
     a = np.arange(15)
-    res = np.int64(0)
+    res = 0
     for ai in a[:i]:
         res += ai
     return res
