+++ tests/pyccel/scripts/import_syntax/from_mod_import.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 from numpy import zeros, shape
 
-#$ header function matmat(float [:,:], float [:,:], float [:,:])
+#$ header function matmat(double [:,:], double [:,:], double [:,:])
 def matmat(a,b,c):
     n, m = shape(a)
     m, p = shape(b)
@@ -16,8 +16,8 @@ if __name__ == '__main__':
     m = 4
     p = 3
 
-    a = zeros((n,m), 'float')
-    b = zeros((m,p), 'float')
+    a = zeros((n,m), 'double')
+    b = zeros((m,p), 'double')
 
     for i in range(0, n):
         for j in range(0, m):
@@ -33,7 +33,7 @@ if __name__ == '__main__':
         print()
     print()
 
-    c = zeros((n,p),'float')
+    c = zeros((n,p),'double')
     matmat(a,b,c)
 
     for i in range(0, n):
