+++ tests/pyccel/scripts/import_syntax/from_mod_import_as.py
@@ -2,7 +2,7 @@
 from numpy import zeros as zilch
 from numpy import shape as form
 
-#$ header function matmat(float [:,:], float [:,:], float [:,:])
+#$ header function matmat(double [:,:], double [:,:], double [:,:])
 def matmat(a,b,c):
     nm = form(a)
     mp = form(b)
@@ -21,8 +21,8 @@ if __name__ == '__main__':
     m = 4
     p = 3
 
-    a = zilch((n,m), 'float')
-    b = zilch((m,p), 'float')
+    a = zilch((n,m), 'double')
+    b = zilch((m,p), 'double')
 
     for i in range(0, n):
         for j in range(0, m):
@@ -38,7 +38,7 @@ if __name__ == '__main__':
         print()
     print()
 
-    c = zilch((n,p),'float')
+    c = zilch((n,p),'double')
     matmat(a,b,c)
 
     for i in range(0, n):
