+++ tests/pyccel/scripts/import_syntax/import_mod_as_user_func.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 
-def fun(xi1 : 'float', xi2 : 'float', xi3 : 'float'):
+def fun(xi1 : 'double', xi2 : 'double', xi3 : 'double'):
     import user_mod as u
     return u.user_func(xi1, xi2, xi3)
