+++ tests/pyccel/scripts/import_syntax/user_mod2.py
@@ -1,4 +1,4 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 
-def user_func(x1 : 'float', x2 : 'float'):
+def user_func(x1 : 'double', x2 : 'double'):
     return x1+x2
